Issues/2-improvised (#79)
* Put team.txt back in main

* updated team.txt

* Create android.yml

Create Android CI workflow

* Fix remaining data classes, add profile screen and edit profile screen (only for display)

* Delete .github/workflows directory

* Refined EditProfileFragment java class

* Refined ProfileFragment java class

* Refined profile screen layout

---------

+++ code/app/src/main/java/com/example/qrquest/ReusableRank.java
@@ -2,7 +2,7 @@ package com.example.qrquest;
 
 /**
  * This interface allows the ArrayList<> of Rank's children classes to be reusable
- * @author: Dang Viet Anh Dinh
+ * @author Dang Viet Anh Dinh
  */
 public interface ReusableRank {
 
@@ -31,6 +31,6 @@ public interface ReusableRank {
      * @param value: the item value
      * @return the rank of the item value or 0 if not found
      */
-    public int getQueryRank(double value);
+    public int getQueryRank(int value);
 
 }
